Split output module into directory
