+++ noline/src/editor/sync_editor.rs
@@ -6,16 +6,13 @@
 //! Use the [`crate::builder::EditorBuilder`] to build an editor.
 #![allow(elided_named_lifetimes)]
 
-use embedded_io::{Read, ReadExactError, Write};
-
+use crate::core::{Line, Prompt};
 use crate::error::NolineError;
-
 use crate::history::{get_history_entries, CircularSlice, History};
 use crate::line_buffer::{Buffer, LineBuffer};
-
-use crate::core::{Line, Prompt};
 use crate::output::{Output, OutputItem};
 use crate::terminal::Terminal;
+use embedded_io::{Read, ReadExactError, Write};
 
 /// Line editor for synchronous IO
 ///
